Added documentation and fixed mutagen metadata bug
All inputted metadata had to be string, but the index was
an integer, so it had to be converted to a string to actually
be able to be placed in the metadata. Also, KeyboardInterrupt
allows the user more options instead of just skipping the video.
Users can now retry the same video, skip to next one, or exit.
Theres probably some small formatting changes, I dont remember.